Add ARCH_ALIGN_UINT64_T_MOD
Note that since uint64_t is defined in stdint_.h which depends on
genarch.c. So we cannot just define a struct with an uint64_t member and
use OFFSET_IN.

Instead, use C11 features to determine the alignment and fall back to
the pointer width, which should work on most platforms.